Fixes after fresh refactoring

+++ frosted/include/cirbuf.h
@@ -15,7 +15,7 @@ int cirbuf_writebytes(struct cirbuf *cb, const uint8_t * bytes, int len);
 
 /* len on success, -1 on fail */
 int cirbuf_readbytes(struct cirbuf *cb, void *bytes, int len);
-int cirbuf_bytesfree(struct cirbuf *cb);
-int cirbuf_bytesinuse(struct cirbuf *cb);
+size_t cirbuf_bytesfree(struct cirbuf *cb);
+size_t cirbuf_bytesinuse(struct cirbuf *cb);
 
 #endif
